fix: poll related fixes
