+++ lib/app/features/feed/providers/ion_connect_entity_with_counters_provider.c.dart
@@ -44,5 +44,12 @@ IonConnectEntity? ionConnectEntityWithCounters(
     ).extensions,
   ]).toString();
 
-  return ref.watch(ionConnectSyncEntityProvider(eventReference: eventReference, search: search));
+  return ref.watch(
+    ionConnectSyncEntityProvider(
+      eventReference: eventReference,
+      search: search,
+      network: network,
+      cache: cache,
+    ),
+  );
 }
